requirements.txt
aiohttp==3.8.1
aiopg==1.3.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
charset-normalizer==2.0.12
click==8.0.4
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
six==1.16.0
SQLAlchemy==1.4.32
sqlalchemy-json==0.5.0
tomli==2.0.1
typing_extensions==4.1.1
yarl==1.7.2
